fix: Update batch_exports.py module
