+++ posthog/temporal/tests/batch_exports/test_batch_exports.py
@@ -1,4 +1,3 @@
-import asyncio
 import datetime as dt
 import json
 import operator
@@ -9,13 +8,10 @@ from django.test import override_settings
 
 from posthog.batch_exports.service import BatchExportModel
 from posthog.temporal.batch_exports.batch_exports import (
-    RecordBatchProducerError,
-    TaskNotDoneError,
     generate_query_ranges,
     get_data_interval,
     iter_model_records,
     iter_records,
-    raise_on_produce_task_failure,
 )
 from posthog.temporal.tests.utils.events import generate_test_events_in_clickhouse
 
@@ -409,48 +405,6 @@ def test_get_data_interval(interval, data_interval_end, expected):
     """Test get_data_interval returns the expected data interval tuple."""
     result = get_data_interval(interval, data_interval_end)
     assert result == expected
-
-
-async def test_raise_on_produce_task_failure_raises_record_batch_producer_error():
-    """Test a `RecordBatchProducerError` is raised with the right cause."""
-    cause = ValueError("Oh no!")
-
-    async def fake_produce_task():
-        raise cause
-
-    task = asyncio.create_task(fake_produce_task())
-    await asyncio.wait([task])
-
-    with pytest.raises(RecordBatchProducerError) as exc_info:
-        await raise_on_produce_task_failure(task)
-
-    assert exc_info.type == RecordBatchProducerError
-    assert exc_info.value.__cause__ == cause
-
-
-async def test_raise_on_produce_task_failure_raises_task_not_done():
-    """Test a `TaskNotDoneError` is raised if we don't let the task start."""
-    cause = ValueError("Oh no!")
-
-    async def fake_produce_task():
-        raise cause
-
-    task = asyncio.create_task(fake_produce_task())
-
-    with pytest.raises(TaskNotDoneError):
-        await raise_on_produce_task_failure(task)
-
-
-async def test_raise_on_produce_task_failure_does_not_raise():
-    """Test nothing is raised if task finished succesfully."""
-
-    async def fake_produce_task():
-        return True
-
-    task = asyncio.create_task(fake_produce_task())
-    await asyncio.wait([task])
-
-    await raise_on_produce_task_failure(task)
 
 
 @pytest.mark.parametrize(
